requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
pymongo==4.6.1
motor==3.3.2
sqlalchemy==2.0.37
databases==0.9.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.0.1
email-validator==2.2.0

# Data Processing
pandas==2.2.0
yfinance==0.2.36
aiohttp==3.9.3

# Development Tools
pyngrok==7.0.0
python-dotenv==1.0.0

# Additional Dependencies
pydantic[email]==2.10.6
pydantic-settings==2.1.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
PyJWT==2.8.0